requirements.txt
# Core dependencies
streamlit>=1.28.1
openai>=1.3.7
python-dotenv>=1.0.0
tiktoken>=0.5.2


# PDF parsing
PyMuPDF>=1.23.0 
PyPDF2>=1.26.0

# Embeddings & similarity
scikit-learn>=1.0.0
numpy>=1.20.0

# Optional
tqdm>=4.65.0  # For progress bars used in preprocessing
